Fix TOP-5767 (OID 3668): Cluster intersection points
Similar to commit 4e8ed63129841e9d2e0fca18cb51d2efede6ceb5
but this time both geometries have a vertex. The vertices are just
above the tolerance, but the intersection points on the source are not.
Previously the determination of the left/rigth side of an intersection
was flawed in this case.