Bugfix: Setting SelectionCursors to null in OneClickToolBase, throws error. This happend in Many contruction tools because the SelectionCursor was set to FirstPhaseCursor, which was null.
We changed the FirstPhaseCursors in ConstructionToolBase to always return SelectionCursors unless overriden by a child.
In OneClickToolBase, we set the backing field for SelectionCursors to it's standard value in OnToolActivatingCoreAsync, to ensure this happens on the GUI thread.
As such we can guarantee that SelectionCursors is never null and cannot be set to null.